requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
atlastk==0.13.2
bcrypt==4.0.1
black==22.12.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dictdiffer==0.9.0
ecdsa==0.18.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pypika-tortoise==0.1.6
python-decouple==3.6
python-jose==3.3.0
pytz==2022.7
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
tortoise==0.1.1
tortoise-orm==0.19.2
typing-extensions==4.4.0
uvicorn==0.20.0
